% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar.R
\name{classificar}
\alias{classificar}
\title{Acrescenta coluna com os tipos das cônicas}
\usage{
classificar(coefs)
}
\arguments{
\item{coefs}{Tibble com 5 colunas: \code{A}, \code{C}, \code{D}, \code{E}, \code{F}}
}
\value{
Tibble original, com coluna \code{tipo} adicionada
}
\description{
Acrescenta coluna com os tipos das cônicas
}
