% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computar_tipo.R
\name{computar_tipo}
\alias{computar_tipo}
\title{Computa o tipo de uma cônica}
\usage{
computar_tipo(A, C, D, E, F, lambda, mu1, mu2)
}
\arguments{
\item{A}{Coeficiente}

\item{C}{Coeficiente}

\item{D}{Coeficiente}

\item{E}{Coeficiente}

\item{F}{Coeficiente}

\item{lambda}{lambda}

\item{mu1}{mu1}

\item{mu2}{mu2}
}
\value{
String com tipo da cônica
}
\description{
Tipos:
\itemize{
\item concorrentes
\item elipse
\item hipérbole
\item paralelas
\item parábola
\item ponto
\item reta
\item vazio
}
}
