% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_conicas_df.R
\name{gerar_conicas_df}
\alias{gerar_conicas_df}
\title{Gera uma tibble com coeficientes de cônicas não rotacionadas}
\usage{
gerar_conicas_df(faixa = -5:5)
}
\arguments{
\item{faixa}{Vetor com valores permitidos para os coeficientes}
}
\value{
Tibble
}
\description{
As colunas da tibble gerada são
\itemize{
\item Os coeficientes de
\itemize{
\item x^2 (\code{A}),
\item y^2 (\code{C}),
\item x (\code{D}),
\item y (\code{E}),
}
\item O termo independente (\code{F})
\item O tipo de cônica (\code{tipo}): vazio, elipse, ponto, paralelas, parábola, reta, hipérbole, ou concorrentes
\item A equação como um string em LaTeX, sem "$" (\code{eq})
}
}
