% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_equacao.R
\name{gerar_equacao}
\alias{gerar_equacao}
\title{Formata a equação de uma cônica}
\usage{
gerar_equacao(A, C, D, E, F)
}
\arguments{
\item{A}{Coeficiente}

\item{C}{Coeficiente}

\item{D}{Coeficiente}

\item{E}{Coeficiente}

\item{F}{Coeficiente}
}
\value{
String com a equação
}
\description{
Formata a equação de uma cônica
}
