% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_lista.R
\name{gerar_lista}
\alias{gerar_lista}
\title{Formata as equações das cônicas}
\usage{
gerar_lista(df, num = TRUE, tex = TRUE, sep = "\\n\\n")
}
\arguments{
\item{df}{Tibble com as cônicas}

\item{num}{Numerar as equações?}

\item{tex}{Formatar as equações em LaTeX?}

\item{sep}{Separador de equações}
}
\value{
String com a lista de equações
}
\description{
Formata as equações das cônicas
}
