% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortear_conicas_um_tipo.R
\name{sortear_conicas_um_tipo}
\alias{sortear_conicas_um_tipo}
\title{Sorteia linhas (de um determinado tipo) da tibble de cônicas}
\usage{
sortear_conicas_um_tipo(df, n_itens, tipo_desejado, seed = NULL)
}
\arguments{
\item{df}{Tibble de cônicas}

\item{n_itens}{Qtde de linhas a sortear}

\item{tipo_desejado}{Tipo desejado}

\item{seed}{Para gerador de números aleatórios}
}
\value{
Tibble com linhas sorteadas
}
\description{
Sorteia linhas (de um determinado tipo) da tibble de cônicas
}
